Time: 22 ms (39.60%), Space: 55.9 MB (36.54%) - LeetHub

+++ 1010-pairs-of-songs-with-total-durations-divisible-by-60/1010-pairs-of-songs-with-total-durations-divisible-by-60.java
@@ -14,7 +14,7 @@ class Solution {
                 long size = map.get(rem);
                 pairs += size * (size - 1) / 2;
             } else {
-                pairs += map.get(rem) * map.getOrDefault(60 - rem, 0);
+                pairs += (long) map.get(rem) * map.getOrDefault(60 - rem, 0);
             }
         }
